refactor: rename assetLoader.ts -> assets.ts
